requirements.txt
pandas
numpy
matplotlib
seaborn
sqlalchemy
psycopg2-binary
python-dotenv
# Optional tools (uncomment if needed)
# pytest
# black
# ruff
